feat(web): explain the subtitle score with a factor breakdown
Move the subtitle search modal out of Modals into its own file, matching
the release search modal's treatment of its ranker score. The flat
Score N badge becomes a hover dropdown listing each scoring factor with
its detail, points, and max, backed by Scoring.Factor from the subtitle
scorer.

Retune the badge thresholds from 150/100 to 100/75. The old numbers were
set against a scale only an OpenSubtitles hash match could climb, so
every relay result rendered ghost grey.

Move the pre-existing subtitle_search_modal shell and state tests out of
modals_test.exs into the new subtitle_modal_test.exs alongside it, since
they exercised the function that moved.
